+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 0.0.4
 
 * Added check for updates on QuPath startup
+* Made pre-release notice less obtrusive
 * Added 'Measure -> Show measurement manager' command to enable measurements to be viewed & (optionally) removed
 * Added 'File -> Revert' command to go back to the last saved version for the current image data
 * Added new 'Add intensity features (experimental)' command. This will eventually replace the Haralick features command (and possibly others), since it offers the same functionality in a much more flexible way.  Furthermore, the new command can handle up to 8 channels of fluorescence data (with arbitrary setting of the min/max values used to calculate the graylevel co-occurrence matrix).